KFIM-56 수지 - 근태관리(세션정보 추가) , 근태체크 (반차, 연차 스케줄러 및 메소드 분리)

+++ DREAMLAND/src/main/java/com/dreamland/prj/service/dayoffService.java
@@ -11,7 +11,7 @@ public interface dayoffService {
  
   void loadDayoffData(Model model, EmployeeDto employee);
   List<AppleaveDto> getDayoffListByYear(int empNo, int year);
-  int calculateRemainDayOff(int totalDayOff, int usedDayOff);
+  double calculateRemainDayOff(double totalDayOff, double usedDayOff);
   public List<Integer> getYearList(int empNo);
 
 }
